Add basic events and an attack envelope
The Conductor now keeps track of events that happen between grain
generations and sends the list of events to the param generator. Param
generators take the event list as one of their arguments and can use
those events to change their state.

As an example, I added an envelop generator with an attack. When an
event occurs it's value starts at zero and rises to the max value.