Add timing control, new envelope
A function is now provided when initializing the Conductor that
controls when grains will be emitted.

A new AHR envelope has been added. It is can be automatically
retriggered, and it can either generate a value, a random value
between it's bias and the envelope, or a random value between bias +/-
envelope. This adds a bit more playability and variation to the
system.